store: cache propose metric and flush in batch.
